Add follow suggestions, refactor follow logic

+++ http/follow.go
@@ -11,35 +11,48 @@ import (
 
 // registerFollowRoutes is a helper for registering all Follow routes.
 func (s *Server) registerFollowRoutes(r *mux.Router) {
-	// Create a new Follow (Follow a user).
-	r.HandleFunc("/follow/{user_id:[0-9]+}", s.requireAuth(s.handleCreateFollow)).Methods("POST")
+	// Get ten users to be suggested to the authed user as potential follows.
+	r.HandleFunc("/follow/suggestions/{user_id:[0-9]+}", s.requireAuth(s.handleGetSuggestions)).Methods("GET")
 
-	// Delete an existing Follow (Unfollow a user).
-	r.HandleFunc("/unfollow/{user_id:[0-9]+}", s.requireAuth(s.handleDeleteFollow)).Methods("DELETE")
+	// Create a new Follow.
+	r.HandleFunc("/follow", s.requireAuth(s.handleCreateFollow)).Methods("POST")
+
+	// Delete an existing Follow.
+	r.HandleFunc("/follow/delete/{id:[0-9]+}", s.requireAuth(s.handleDeleteFollow)).Methods("DELETE")
 }
 
-// handleCreateFollow handles the route "POST /follow/:user_id".
-// It reads the ID of the user to be followed from the url and creates
-// a new Follow record in the database.
+func (s *Server) handleGetSuggestions(w http.ResponseWriter, r *http.Request) {
+	// Parse the User ID from the url.
+	userId, err := strconv.Atoi(mux.Vars(r)["user_id"])
+	if userId <=0 || err != nil{
+		errs.ReturnError(w, r, errs.Errorf(errs.EINVALID, "Invalid Id format."))
+		return
+	}
+	suggestions := s.fs.SuggestFollows(userId)
+	w.WriteHeader(http.StatusOK)
+	if err := json.NewEncoder(w).Encode(suggestions); err != nil {
+		errs.LogError(r, err)
+		return
+	}
+}
+
+// handleCreateFollow handles the route "POST /follow".
+// It reads the followed_id from the json body, gets the authed user's id from context,
+// sets their id as the follower_id, and creates a new Follow record in the database.
 func (s *Server) handleCreateFollow(w http.ResponseWriter, r *http.Request) {
-	// Parse the ID of the user to be followed from the url.
-	followedId, err := strconv.Atoi(mux.Vars(r)["user_id"])
-	if err != nil {
-		errs.ReturnError(w, r, errs.Errorf(errs.EINVALID, "Invalid user id."))
+	// Parse the request's json body into a Follow object. It only contains the followed_id.
+	var follow domain.Follow
+	if err := json.NewDecoder(r.Body).Decode(&follow); err != nil {
+		errs.ReturnError(w, r, errs.Errorf(errs.EINVALID, "Invalid json body."))
 		return
 	}
 
-	// Get the authed user form the request context.
+	// Get the authed user's ID and set it as the new Follow's FollowerID.
 	user := s.getUserFromContext(r.Context())
-
-	// Create a Follow object.
-	follow := &domain.Follow{
-		FollowedID: followedId,
-		FollowerID: user.ID,
-	}
+	follow.FollowerID = user.ID
 
 	// Create a new Follow database record.
-	err = s.fs.Create(follow)
+	err := s.fs.Create(&follow)
 	if err != nil {
 		errs.ReturnError(w, r, err)
 		return
@@ -53,33 +66,38 @@ func (s *Server) handleCreateFollow(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-// handleDeleteFollow handles the route "DELETE /unfollow/:user_id".
-// It reads the ID of the followed user from the url and permanently deletes
-// the respective Follow record from the database.
+// handleDeleteFollow handles the route "DELETE /follow/delete/:id".
+// It reads id from the url and permanently deletes the respective follow record from the database.
 func (s *Server) handleDeleteFollow(w http.ResponseWriter, r *http.Request) {
-	// Parse the ID of the followed user from the url.
-	followedId, err := strconv.Atoi(mux.Vars(r)["user_id"])
+	// Parse the follow ID from the url.
+	id, err := strconv.Atoi(mux.Vars(r)["id"])
 	if err != nil {
-		errs.ReturnError(w, r, errs.Errorf(errs.EINVALID, "Invalid user id."))
+		errs.ReturnError(w, r, errs.Errorf(errs.EINVALID, "Invalid Id format."))
 		return
 	}
 
-	// Get the authed user form the request context.
-	user := s.getUserFromContext(r.Context())
-
-	// Create a Follow object.
-	follow := &domain.Follow{
-		FollowedID: followedId,
-		FollowerID: user.ID,
+	// Fetch the follow from the database.
+	follow, err := s.fs.ByID(id)
+	if err != nil {
+		errs.ReturnError(w, r, err)
+		return
 	}
 
-	// Permanently delete the Follow database record.
+	// Check if the follow belongs to the authed user.
+	user := s.getUserFromContext(r.Context())
+	if follow.FollowerID != user.ID {
+		errs.ReturnError(w, r, errs.Errorf(errs.EUNAUTHORIZED, "You are not allowed to delete this follow."))
+		return
+	}
+
+	// Soft-delete the follow.
 	err = s.fs.Delete(follow)
 	if err != nil {
 		errs.ReturnError(w, r, err)
 		return
 	}
 
-	// Return status code 204 to indicate successful deletion.
+	// Return the soft-deleted follow.
+	// TODO: Only return this on successful deletion
 	w.WriteHeader(http.StatusNoContent)
 }
